Fix build & move native dll into different location - now at lib/native
