+++ exonum-java-binding-core/rust/src/utils/jni.rs
@@ -62,22 +62,3 @@ pub fn get_exception_stack_trace(_env: &JNIEnv, exception: JObject) -> JniResult
     //Ok(stack.join(""))
     Ok(String::new())
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    const FOO: &str = "foo";
-    const BAR: &str = "bar";
-    const BAZ: &str = "baz";
-
-    #[cfg(windows)]
-    const FOO_BAR_BAZ: &str = "foo;bar;baz";
-    #[cfg(not(windows))]
-    const FOO_BAR_BAZ: &str = "foo:bar:baz";
-
-    #[test]
-    fn join_paths_preserves_order() {
-        let result = join_paths(&[FOO, BAR, BAZ]);
-        assert_eq!(result, FOO_BAR_BAZ);
-    }
-}
